requirements.txt
loguru==0.7.0
python-telegram-bot[job-queue]==20.3
SQLAlchemy==2.0.14